refactor(modal): config env helper 분리
